+++ tests/testthat/test-SDA_pmgroupname.R
@@ -5,4 +5,6 @@ test_that("get_SDA_pmgroupname works", {
 
   expect_equal(nrow(get_SDA_pmgroupname(areasymbols = c("CA077", "CA630"))), 292)
   expect_equal(nrow(get_SDA_pmgroupname(mukeys = c(461994, 461995), simplify = FALSE)), 2)
+  expect_equal(nrow(get_SDA_pmgroupname(mukeys = c(461994, 461995), simplify = FALSE, method = "none")), 7)
+  expect_equal(nrow(get_SDA_pmgroupname(mukeys = c(461994, 461995), simplify = FALSE, method = "dominant condition")), 2)
 })